Destruction is the mother of creation!
